[#71] MainCoordinator 통합
